fix: PR review — validate addresses, bracket IPv6 URLs, move tests
- address() now rejects empty host/port
- IPv6 literals bracketed in URL construction (http://[::1]:9000/...)
- Rotation unit tests moved from sender/http.rs to tests/http.rs
- Make conf and sender modules pub(crate) for test access
